增加创建用户逻辑

+++ priest-common/src/main/java/com/little/g/common/enums/DeviceTypeEnum.java
@@ -11,16 +11,16 @@ import java.util.Map;
  */
 public enum DeviceTypeEnum {
 
-    MOBILE(1),//手机
-    PAD(2),//平板电脑
-    PC(3),//PC机
-    TV(4),//电视
-    H5(6);//HTML5页面
+    MOBILE((byte)1),//手机
+    PAD((byte)2),//平板电脑
+    PC((byte)3),//PC机
+    TV((byte)4),//电视
+    H5((byte)6);//HTML5页面
 
-    public int value;
+    public Byte value;
 
     //数字与字符串映射字典表
-    public static Map<Integer, DeviceTypeEnum> TYPE_MAPPING_DICT = new LinkedHashMap<>();
+    public static Map<Byte, DeviceTypeEnum> TYPE_MAPPING_DICT = new LinkedHashMap<>();
 
     static {
         TYPE_MAPPING_DICT.put(MOBILE.getValue(), MOBILE);
@@ -30,7 +30,7 @@ public enum DeviceTypeEnum {
         TYPE_MAPPING_DICT.put(H5.getValue(), H5);
     }
 
-    public static boolean checkKeyIsExist(Integer code) {
+    public static boolean checkKeyIsExist(Byte code) {
         if (TYPE_MAPPING_DICT.containsKey(code)) {
             return true;
         }
@@ -38,15 +38,15 @@ public enum DeviceTypeEnum {
     }
 
 
-    public static DeviceTypeEnum getDeviceTypeEnum(int value) {
+    public static DeviceTypeEnum getDeviceTypeEnum(Byte value) {
         return TYPE_MAPPING_DICT.get(value);
     }
 
-    DeviceTypeEnum(int value) {
+    DeviceTypeEnum(Byte value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public Byte getValue() {
         return value;
     }
 }
